continue moving
