+++ social/routes.ts
@@ -1,0 +1,1265 @@
+import type { BmapTx } from 'bmapjs';
+import { Elysia, t } from 'elysia';
+import { InternalServerError, error } from 'elysia';
+import * as mongo from 'mongodb';
+import type { ChangeStream, ChangeStreamDocument, ChangeStreamInsertDocument, Db } from 'mongodb';
+import { getBAPIdByAddress } from '../bap.js';
+import type { BapIdentity } from '../bap.js';
+import { normalize } from '../bmap.js';
+import { client, readFromRedis, saveToRedis } from '../cache.js';
+import type { CacheError, CacheSigner, CacheValue } from '../cache.js';
+import { getDbo } from '../db.js';
+import { getDirectMessages, watchAllMessages, watchDirectMessages } from '../queries/messages.js';
+import { ChannelResponseSchema, channelsEndpointDetail } from './swagger/channels.js';
+import type { ChannelInfo } from './swagger/channels.js';
+import type { Friend, FriendshipResponse, RelationshipState } from './swagger/friend.js';
+import { FriendResponseSchema, friendEndpointDetail } from './swagger/friend.js';
+import { IdentityResponseSchema, identityEndpointDetail } from './swagger/identity.js';
+import type {
+  Identity,
+  SigmaIdentityAPIResponse,
+  SigmaIdentityResult,
+} from './swagger/identity.js';
+import type { LikeRequest, LikeResponse, Reaction, Reactions } from './swagger/likes.js';
+import { LikeRequestSchema, LikeResponseSchema } from './swagger/likes.js';
+import type { ChannelMessage, Message } from './swagger/messages.js';
+import {
+  ChannelMessageSchema,
+  DMResponseSchema,
+  MessageListenParams,
+  MessageQuery,
+  channelMessagesEndpointDetail,
+  messageListenEndpointDetail,
+} from './swagger/messages.js';
+
+function sigmaIdentityToBapIdentity(result: SigmaIdentityResult): BapIdentity {
+  return {
+    idKey: result.idKey,
+    rootAddress: result.rootAddress,
+    currentAddress: result.currentAddress,
+    addresses: result.addresses,
+    identity: result.identity,
+    identityTxId: result.identityTxId,
+    block: result.block,
+    timestamp: result.timestamp,
+    valid: result.valid,
+  };
+}
+
+async function fetchBapIdentityData(bapId: string): Promise<BapIdentity> {
+  const cacheKey = `sigmaIdentity-${bapId}`;
+  const cached = await readFromRedis<CacheValue>(cacheKey);
+  if (cached?.type === 'signer') {
+    return cached.value;
+  }
+
+  const url = 'https://api.sigmaidentity.com/v1/identity/get';
+  const resp = await fetch(url, {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({ idKey: bapId }),
+  });
+
+  if (!resp.ok) {
+    const text = await resp.text();
+    throw new Error(`Failed to fetch identity data. Status: ${resp.status}, Body: ${text}`);
+  }
+
+  const data: SigmaIdentityAPIResponse = await resp.json();
+  if (data.status !== 'OK' || !data.result || data.error) {
+    throw new Error(
+      `Sigma Identity returned invalid data for ${bapId}: ${data.error || 'Unknown error'}`
+    );
+  }
+
+  const bapIdentity = sigmaIdentityToBapIdentity(data.result);
+
+  await saveToRedis<CacheValue>(cacheKey, {
+    type: 'signer',
+    value: bapIdentity,
+  });
+
+  return bapIdentity;
+}
+
+async function fetchAllFriendsAndUnfriends(
+  bapId: string
+): Promise<{ allDocs: BmapTx[]; ownedAddresses: Set<string> }> {
+  console.log('\n=== fetchAllFriendsAndUnfriends ===');
+  console.log('BAP ID:', bapId);
+
+  const dbo = await getDbo();
+
+  const idData = await fetchBapIdentityData(bapId);
+  if (!idData || !idData.addresses) {
+    throw new Error(`No identity found for ${bapId}`);
+  }
+
+  const ownedAddresses = new Set<string>(idData.addresses.map((a) => a.address));
+  console.log('Owned addresses:', [...ownedAddresses]);
+
+  // Get incoming friend requests (where this BAP ID is the target)
+  const incomingFriends = (await dbo
+    .collection('friend')
+    .find({ 'MAP.type': 'friend', 'MAP.bapID': bapId })
+    .toArray()) as unknown as BmapTx[];
+
+  console.log('Incoming friends count:', incomingFriends.length);
+  console.log(
+    'Incoming friends:',
+    JSON.stringify(
+      incomingFriends.map((f) => ({
+        txid: f.tx?.h,
+        bapID: f.MAP?.[0]?.bapID,
+        address: f.AIP?.[0]?.algorithm_signing_component || f.AIP?.[0]?.address,
+      })),
+      null,
+      2
+    )
+  );
+
+  // Get outgoing friend requests (where this BAP ID's addresses are the source)
+  const outgoingFriends = (await dbo
+    .collection('friend')
+    .find({
+      'MAP.type': 'friend',
+      $or: [
+        { 'AIP.algorithm_signing_component': { $in: [...ownedAddresses] } },
+        { 'AIP.address': { $in: [...ownedAddresses] } },
+      ],
+    })
+    .toArray()) as unknown as BmapTx[];
+
+  // Try to get unfriend documents if the collection exists
+  let incomingUnfriends: BmapTx[] = [];
+  let outgoingUnfriends: BmapTx[] = [];
+
+  try {
+    const collections = await dbo.listCollections().toArray();
+    const hasUnfriendCollection = collections.some((c) => c.name === 'unfriend');
+
+    if (hasUnfriendCollection) {
+      incomingUnfriends = (await dbo
+        .collection('unfriend')
+        .find({ 'MAP.type': 'unfriend', 'MAP.bapID': bapId })
+        .toArray()) as unknown as BmapTx[];
+
+      outgoingUnfriends = (await dbo
+        .collection('unfriend')
+        .find({
+          'MAP.type': 'unfriend',
+          $or: [
+            { 'AIP.algorithm_signing_component': { $in: [...ownedAddresses] } },
+            { 'AIP.address': { $in: [...ownedAddresses] } },
+          ],
+        })
+        .toArray()) as unknown as BmapTx[];
+    }
+  } catch (error) {
+    console.warn('Failed to query unfriend collection:', error);
+  }
+
+  console.log('Outgoing friends count:', outgoingFriends.length);
+  console.log(
+    'Outgoing friends:',
+    JSON.stringify(
+      outgoingFriends.map((f) => ({
+        txid: f.tx?.h,
+        bapID: f.MAP?.[0]?.bapID,
+        address: f.AIP?.[0]?.algorithm_signing_component || f.AIP?.[0]?.address,
+      })),
+      null,
+      2
+    )
+  );
+
+  const allDocs = [
+    ...incomingFriends,
+    ...incomingUnfriends,
+    ...outgoingFriends,
+    ...outgoingUnfriends,
+  ];
+  allDocs.sort((a, b) => (a.blk?.i ?? 0) - (b.blk?.i ?? 0));
+
+  console.log('Total documents:', allDocs.length);
+  return { allDocs, ownedAddresses };
+}
+
+async function processRelationships(
+  bapId: string,
+  docs: BmapTx[],
+  ownedAddresses: Set<string>
+): Promise<FriendshipResponse> {
+  console.log('\n=== processRelationships ===');
+  console.log('Processing relationships for BAP ID:', bapId);
+  console.log('Number of documents:', docs.length);
+  console.log('Owned addresses:', [...ownedAddresses]);
+
+  const relationships = new Map<string, RelationshipState>();
+
+  async function getRequestorBapId(doc: BmapTx): Promise<string | null> {
+    // Check all possible address fields
+    const address = doc?.AIP?.[0]?.algorithm_signing_component || doc?.AIP?.[0]?.address;
+    if (!address) {
+      console.log('No address found in document:', doc.tx?.h);
+      return null;
+    }
+
+    if (ownedAddresses.has(address)) {
+      console.log('Address matches owned address:', address);
+      return bapId;
+    }
+
+    console.log('Looking up BAP ID for address:', address);
+    const otherIdentity = await getBAPIdByAddress(address);
+    if (!otherIdentity) {
+      console.log('No identity found for address:', address);
+      return null;
+    }
+    console.log('Found BAP ID for address:', otherIdentity.idKey);
+    return otherIdentity.idKey;
+  }
+
+  const requestors = await Promise.all(docs.map((doc) => getRequestorBapId(doc)));
+  console.log('Resolved requestors:', requestors);
+
+  for (let i = 0; i < docs.length; i++) {
+    const doc = docs[i];
+    const reqBap = requestors[i];
+    const tgtBap = doc?.MAP?.[0]?.bapID;
+    const publicKey = doc?.MAP?.[0]?.publicKey;
+
+    console.log('\nProcessing document:', doc.tx?.h);
+    console.log('Requestor BAP:', reqBap);
+    console.log('Target BAP:', tgtBap);
+
+    if (!reqBap || !tgtBap || !Array.isArray(doc.MAP)) {
+      console.log('Skipping document - missing required fields');
+      continue;
+    }
+
+    const otherBapId = reqBap === bapId ? tgtBap : reqBap;
+    console.log('Other BAP ID:', otherBapId);
+
+    if (otherBapId && typeof otherBapId === 'string' && !relationships.has(otherBapId)) {
+      console.log('Creating new relationship for:', otherBapId);
+      relationships.set(otherBapId, { fromMe: false, fromThem: false, unfriended: false });
+    }
+
+    const rel = relationships.get(typeof otherBapId === 'string' ? otherBapId : '');
+    if (!rel) {
+      console.log('No relationship found for:', otherBapId);
+      continue;
+    }
+
+    const isFriend = doc?.MAP?.[0]?.type === 'friend';
+    const isUnfriend = doc?.MAP?.[0]?.type === 'unfriend';
+    const isFromMe = reqBap === bapId;
+
+    console.log('Document type:', isFriend ? 'friend' : isUnfriend ? 'unfriend' : 'unknown');
+    console.log('Is from me:', isFromMe);
+
+    if (isUnfriend) {
+      console.log('Processing unfriend');
+      rel.unfriended = true;
+      rel.fromMe = false;
+      rel.fromThem = false;
+    } else if (isFriend) {
+      console.log('Processing friend');
+      if (rel.unfriended) {
+        rel.unfriended = false;
+      }
+      if (isFromMe) {
+        rel.fromMe = true;
+        rel.mePublicKey = publicKey;
+      } else {
+        rel.fromThem = true;
+        rel.themPublicKey = publicKey;
+      }
+    }
+
+    console.log(
+      'Updated relationship:',
+      JSON.stringify({
+        otherBapId,
+        fromMe: rel.fromMe,
+        fromThem: rel.fromThem,
+        unfriended: rel.unfriended,
+      })
+    );
+  }
+
+  const friends: Friend[] = [];
+  const incoming: string[] = [];
+  const outgoing: string[] = [];
+
+  console.log('\nFinal relationships:');
+  for (const [other, rel] of relationships.entries()) {
+    console.log('Processing final relationship:', other, JSON.stringify(rel));
+
+    if (rel.unfriended) {
+      console.log('Skipping unfriended relationship:', other);
+      continue;
+    }
+    if (rel.fromMe && rel.fromThem) {
+      console.log('Adding mutual friend:', other);
+      friends.push({
+        bapID: other,
+        mePublicKey: rel.mePublicKey || '',
+        themPublicKey: rel.themPublicKey || '',
+      });
+    } else if (rel.fromMe && !rel.fromThem) {
+      console.log('Adding outgoing friend:', other);
+      outgoing.push(other);
+    } else if (!rel.fromMe && rel.fromThem) {
+      console.log('Adding incoming friend:', other);
+      incoming.push(other);
+    }
+  }
+
+  console.log('\nFinal results:');
+  console.log('Friends:', friends);
+  console.log('Incoming:', incoming);
+  console.log('Outgoing:', outgoing);
+
+  return { friends, incoming, outgoing };
+}
+
+// Validation helper for signer data
+function validateSignerData(signer: BapIdentity): { isValid: boolean; errors: string[] } {
+  const errors: string[] = [];
+
+  if (!signer.idKey) errors.push('Missing idKey');
+  if (!signer.currentAddress) errors.push('Missing currentAddress');
+  if (!signer.rootAddress) errors.push('Missing rootAddress');
+  if (!signer.addresses || !signer.addresses.length) errors.push('Missing addresses');
+
+  return {
+    isValid: errors.length === 0,
+    errors,
+  };
+}
+
+// Helper function to merge new signers into cache
+async function updateSignerCache(newSigners: BapIdentity[]): Promise<void> {
+  for (const signer of newSigners) {
+    const signerKey = `signer-${signer.currentAddress}`;
+    await saveToRedis<CacheValue>(signerKey, {
+      type: 'signer',
+      value: signer,
+    });
+  }
+  // Clear the identities cache to force a refresh with new signers
+  await client.del('identities');
+}
+
+// Helper function to resolve signers from messages
+async function resolveSigners(messages: Message[]): Promise<BapIdentity[]> {
+  const signerAddresses = new Set<string>();
+
+  for (const msg of messages) {
+    if (msg.AIP && Array.isArray(msg.AIP)) {
+      for (const aip of msg.AIP) {
+        const address = aip.algorithm_signing_component || aip.address;
+        if (address) {
+          signerAddresses.add(address);
+        }
+      }
+    }
+  }
+
+  const signers = await Promise.all(
+    Array.from(signerAddresses).map(async (address) => {
+      try {
+        const identity = await getBAPIdByAddress(address);
+        if (identity) {
+          // Update the signer cache with this identity
+          const signerKey = `signer-${address}`;
+          await saveToRedis<CacheValue>(signerKey, {
+            type: 'signer',
+            value: identity,
+          });
+          return identity;
+        }
+      } catch (error) {
+        console.error(`Failed to resolve signer for address ${address}:`, error);
+      }
+      return null;
+    })
+  );
+
+  const validSigners = signers.filter((s): s is BapIdentity => s !== null);
+
+  // Update the cache with new signers
+  await updateSignerCache(validSigners);
+
+  return validSigners;
+}
+
+async function getChannels() {
+  try {
+    const cacheKey = 'channels';
+    const cached = await readFromRedis<CacheValue>(cacheKey);
+
+    console.log('channels cache key', cacheKey);
+    if (cached?.type === 'channels') {
+      console.log('Cache hit for channels');
+      return cached.value;
+    }
+
+    console.log('Cache miss for channels');
+    const db = await getDbo();
+
+    try {
+      const pipeline = [
+        {
+          $match: {
+            'MAP.channel': { $exists: true, $ne: '' },
+          },
+        },
+        {
+          $unwind: '$MAP',
+        },
+        {
+          $unwind: '$B',
+        },
+        {
+          $group: {
+            _id: '$MAP.channel',
+            channel: { $first: '$MAP.channel' },
+            creator: { $first: { $ifNull: ['$MAP.paymail', null] } },
+            last_message: { $last: { $ifNull: ['$B.Data.utf8', null] } },
+            last_message_time: { $max: '$blk.t' },
+            messages: { $sum: 1 },
+          },
+        },
+        {
+          $sort: { last_message_time: -1 },
+        },
+        {
+          $limit: 100,
+        },
+      ];
+
+      console.log('Executing MongoDB aggregation');
+      const results = await db.collection('message').aggregate(pipeline).toArray();
+      console.log('Aggregation results:', results.length);
+
+      const channels = results.map((r) => ({
+        channel: r.channel,
+        creator: r.creator || null,
+        last_message: r.last_message || null,
+        last_message_time: r.last_message_time,
+        messages: r.messages,
+      }));
+
+      await saveToRedis<CacheValue>(cacheKey, {
+        type: 'channels',
+        value: channels,
+      });
+
+      return channels;
+    } catch (dbError) {
+      console.error('MongoDB operation failed:', dbError);
+      throw new Error('Failed to fetch channels');
+    }
+  } catch (error) {
+    console.error('getChannels error:', error);
+    throw error;
+  }
+}
+
+// Helper to process likes with better error handling and logging
+async function processLikes(
+  likes: Reaction[]
+): Promise<{ signerIds: string[]; signers: BapIdentity[] }> {
+  console.log('Processing likes:', likes.length);
+
+  // Get unique signer addresses with validation
+  const signerAddresses = new Set<string>();
+  const invalidLikes: string[] = [];
+
+  for (const like of likes) {
+    if (!Array.isArray(like.AIP)) {
+      console.warn('Invalid like document - missing AIP array:', like.tx?.h);
+      invalidLikes.push(like.tx?.h);
+      continue;
+    }
+
+    for (const aip of like.AIP) {
+      if (!aip.algorithm_signing_component) {
+        console.warn('Invalid AIP entry - missing algorithm_signing_component:', like.tx?.h);
+        continue;
+      }
+      signerAddresses.add(aip.algorithm_signing_component);
+    }
+  }
+
+  if (invalidLikes.length > 0) {
+    console.warn('Found invalid like documents:', invalidLikes);
+  }
+
+  console.log('Found unique signer addresses:', signerAddresses.size);
+
+  // Fetch and validate signer identities
+  const signerIds = Array.from(signerAddresses);
+  const signerResults = await Promise.all(
+    signerIds.map(async (address) => {
+      const signerCacheKey = `signer-${address}`;
+      const cachedSigner = await readFromRedis<CacheValue>(signerCacheKey);
+
+      if (cachedSigner?.type === 'signer' && cachedSigner.value) {
+        const validation = validateSignerData(cachedSigner.value);
+        if (!validation.isValid) {
+          console.warn(
+            'Invalid cached signer data for address:',
+            address,
+            'Errors:',
+            validation.errors
+          );
+          return null;
+        }
+        return cachedSigner.value;
+      }
+
+      try {
+        const identity = await getBAPIdByAddress(address);
+        if (identity) {
+          const validation = validateSignerData(identity);
+          if (!validation.isValid) {
+            console.warn(
+              'Invalid fetched signer data for address:',
+              address,
+              'Errors:',
+              validation.errors
+            );
+            return null;
+          }
+
+          await saveToRedis<CacheValue>(signerCacheKey, {
+            type: 'signer',
+            value: identity,
+          });
+          return identity;
+        }
+      } catch (error) {
+        console.error(`Failed to fetch identity for address ${address}:`, error);
+      }
+      return null;
+    })
+  );
+
+  const validSigners = signerResults.filter((s): s is BapIdentity => s !== null);
+  console.log('Successfully processed signers:', validSigners.length);
+
+  return {
+    signerIds: validSigners.map((s) => s.idKey),
+    signers: validSigners,
+  };
+}
+
+const ChannelParams = t.Object({
+  channelId: t.String(),
+});
+
+export const socialRoutes = new Elysia()
+  .get(
+    '/channels',
+    async ({ set }) => {
+      try {
+        const channels = await getChannels();
+        Object.assign(set.headers, {
+          'Cache-Control': 'public, max-age=60',
+        });
+        return channels;
+      } catch (err) {
+        console.error('Route handler error:', err);
+        set.status = 500;
+        Object.assign(set.headers, {
+          'Cache-Control': 'no-store',
+        });
+        return {
+          code: 'INTERNAL_SERVER_ERROR',
+          message: 'Failed to fetch channels',
+        };
+      }
+    },
+    {
+      detail: channelsEndpointDetail,
+      response: t.Union([
+        ChannelResponseSchema,
+        t.Object({
+          code: t.String(),
+          message: t.String(),
+        }),
+      ]),
+    }
+  )
+  .get(
+    '/channels/:channelId/messages',
+    async ({ params, query, set }) => {
+      try {
+        const { channelId } = params;
+        if (!channelId) {
+          throw new Error('Missing channel ID');
+        }
+
+        const decodedChannelId = decodeURIComponent(channelId);
+
+        const page = query.page ? Number.parseInt(query.page, 10) : 1;
+        const limit = query.limit ? Number.parseInt(query.limit, 10) : 100;
+
+        if (Number.isNaN(page) || page < 1) {
+          throw new Error('Invalid page parameter');
+        }
+
+        if (Number.isNaN(limit) || limit < 1 || limit > 1000) {
+          throw new Error('Invalid limit parameter');
+        }
+
+        const skip = (page - 1) * limit;
+
+        const cacheKey = `messages:${decodedChannelId}:${page}:${limit}`;
+        const cached = await readFromRedis<CacheValue>(cacheKey);
+
+        if (cached?.type === 'messages') {
+          console.log('Cache hit for messages:', cacheKey);
+          Object.assign(set.headers, {
+            'Cache-Control': 'public, max-age=60',
+          });
+          const response: ChannelMessage = {
+            ...cached.value,
+            channel: channelId,
+            signers: cached.value.signers || [],
+          };
+          return response;
+        }
+
+        console.log('Cache miss for messages:', cacheKey);
+        const db = await getDbo();
+
+        const queryObj = {
+          'MAP.type': 'message',
+          'MAP.channel': decodedChannelId,
+        };
+
+        const col = db.collection('message');
+
+        const count = await col.countDocuments(queryObj);
+
+        const results = (await col
+          .find(queryObj)
+          .sort({ 'blk.t': -1 })
+          .skip(skip)
+          .limit(limit)
+          .project({ _id: 0 })
+          .toArray()) as Message[];
+
+        // Normalize and validate each message
+        const validatedResults = results.map((msg) => ({
+          ...msg,
+          tx: { h: msg.tx?.h || '' },
+          blk: { i: msg.blk?.i || 0, t: msg.blk?.t || 0 },
+          MAP: msg.MAP?.map((m) => ({
+            app: m.app || '',
+            type: m.type || '',
+            channel: m.channel || '',
+            paymail: m.paymail || '',
+          })) || [
+            {
+              app: '',
+              type: '',
+              channel: '',
+              paymail: '',
+            },
+          ],
+          B: msg.B?.map((b) => ({
+            encoding: b?.encoding || '',
+            Data: {
+              utf8: b.Data?.utf8 || '',
+              data: b.Data?.data,
+            },
+          })) || [
+            {
+              encoding: '',
+              Data: {
+                utf8: '',
+                data: '',
+              },
+            },
+          ],
+        }));
+
+        // Initialize empty signers array with proper type
+        let signers: BapIdentity[] = [];
+
+        // Only try to resolve signers if there are messages with AIP data
+        const messagesWithAIP = results.filter((msg) => msg.AIP && msg.AIP.length > 0);
+        if (messagesWithAIP.length > 0) {
+          try {
+            signers = await resolveSigners(messagesWithAIP);
+            console.log(`Resolved ${signers.length} signers`);
+          } catch (error) {
+            console.error('Error resolving signers:', error);
+            // Don't throw - continue with empty signers array
+          }
+        } else {
+          console.log('No messages with AIP data found');
+        }
+
+        // Ensure signers array is properly initialized with all required fields
+        const validatedSigners: BapIdentity[] = signers.map((s) => ({
+          ...s,
+          identityTxId: s.identityTxId || '',
+          identity:
+            typeof s.identity === 'string'
+              ? s.identity
+              : typeof s.identity === 'object'
+                ? s.identity
+                : JSON.stringify(s.identity || {}),
+        }));
+
+        const response: ChannelMessage = {
+          channel: channelId,
+          page,
+          limit,
+          count,
+          results: validatedResults,
+          signers: validatedSigners,
+        };
+
+        await saveToRedis<CacheValue>(cacheKey, {
+          type: 'messages',
+          value: response,
+        });
+
+        Object.assign(set.headers, {
+          'Cache-Control': 'public, max-age=60',
+        });
+        return response;
+      } catch (error: unknown) {
+        console.error('Error fetching messages:', error);
+        set.status = 500;
+        // Return a properly structured response with empty arrays
+        const errorResponse: ChannelMessage = {
+          channel: params.channelId || '',
+          page: 1,
+          limit: 100,
+          count: 0,
+          results: [
+            {
+              tx: { h: '' },
+              blk: { i: 0, t: 0 },
+              MAP: [{ app: '', type: '', channel: '', paymail: '' }],
+              B: [{ encoding: '', Data: { utf8: '', data: '' } }],
+            },
+          ],
+          signers: [],
+        };
+        return errorResponse;
+      }
+    },
+    {
+      params: ChannelParams,
+      query: MessageQuery,
+      response: ChannelMessageSchema,
+      detail: channelMessagesEndpointDetail,
+    }
+  )
+  .post(
+    '/likes',
+    async ({ body }) => {
+      try {
+        const request = body as LikeRequest;
+        if (!request.txids && !request.messageIds) {
+          throw new Error('Must provide either txids or messageIds');
+        }
+
+        const db = await getDbo();
+        const results: LikeResponse[] = [];
+
+        if (request.txids) {
+          for (const txid of request.txids) {
+            const likes = (await db
+              .collection('like')
+              .find({
+                'MAP.type': 'like',
+                'MAP.tx': txid,
+              })
+              .toArray()) as unknown as Reaction[];
+
+            const { signers } = await processLikes(likes);
+
+            results.push({
+              txid,
+              likes,
+              total: likes.length,
+              signers,
+            });
+          }
+        }
+
+        if (request.messageIds) {
+          for (const messageId of request.messageIds) {
+            const likes = (await db
+              .collection('like')
+              .find({
+                'MAP.type': 'like',
+                'MAP.messageID': messageId,
+              })
+              .toArray()) as unknown as Reaction[];
+
+            const { signers } = await processLikes(likes);
+
+            results.push({
+              txid: messageId,
+              likes,
+              total: likes.length,
+              signers,
+            });
+          }
+        }
+
+        return results;
+      } catch (error: unknown) {
+        console.error('Error processing likes:', error);
+        const message = error instanceof Error ? error.message : String(error);
+        throw new Error(`Failed to process likes: ${message}`);
+      }
+    },
+    {
+      body: LikeRequestSchema,
+      response: LikeResponseSchema,
+      detail: {
+        tags: ['social'],
+        description: 'Get likes for transactions or messages',
+        summary: 'Get likes',
+        requestBody: {
+          description: 'Transaction IDs or Message IDs to get likes for',
+          required: true,
+          content: {
+            'application/json': {
+              schema: {
+                type: 'object',
+                properties: {
+                  txids: {
+                    type: 'array',
+                    items: { type: 'string' },
+                    description: 'List of transaction IDs',
+                  },
+                  messageIds: {
+                    type: 'array',
+                    items: { type: 'string' },
+                    description: 'List of message IDs',
+                  },
+                },
+              },
+            },
+          },
+        },
+        responses: {
+          200: {
+            description: 'Likes with signer information',
+            content: {
+              'application/json': {
+                schema: {
+                  type: 'array',
+                  items: {
+                    type: 'object',
+                    properties: {
+                      txid: { type: 'string' },
+                      likes: {
+                        type: 'array',
+                        items: {
+                          type: 'object',
+                          properties: {
+                            tx: { type: 'object', properties: { h: { type: 'string' } } },
+                            blk: {
+                              type: 'object',
+                              properties: {
+                                i: { type: 'number' },
+                                t: { type: 'number' },
+                              },
+                            },
+                            MAP: {
+                              type: 'array',
+                              items: {
+                                type: 'object',
+                                properties: {
+                                  type: { type: 'string' },
+                                  tx: { type: 'string' },
+                                  messageID: { type: 'string' },
+                                  emoji: { type: 'string' },
+                                },
+                              },
+                            },
+                          },
+                        },
+                      },
+                      total: { type: 'number' },
+                      signers: {
+                        type: 'array',
+                        items: {
+                          $ref: '#/components/schemas/BapIdentity',
+                        },
+                      },
+                    },
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    }
+  )
+  .get(
+    '/friend/:bapId',
+    async ({ params, set }) => {
+      try {
+        const { bapId } = params;
+        if (!bapId) {
+          set.status = 400;
+          return {
+            friends: [],
+            incoming: [],
+            outgoing: [],
+          };
+        }
+
+        try {
+          // First verify the BAP ID exists by trying to fetch its identity
+          const identity = await fetchBapIdentityData(bapId);
+          if (!identity) {
+            set.status = 404;
+            return {
+              friends: [],
+              incoming: [],
+              outgoing: [],
+            };
+          }
+
+          const { allDocs, ownedAddresses } = await fetchAllFriendsAndUnfriends(bapId);
+          return processRelationships(bapId, allDocs, ownedAddresses);
+        } catch (error) {
+          // If we get an error fetching the BAP identity, return 404
+          console.error('Error fetching BAP identity:', error);
+          set.status = 404;
+          return {
+            friends: [],
+            incoming: [],
+            outgoing: [],
+          };
+        }
+      } catch (error: unknown) {
+        console.error('Error processing friend request:', error);
+        set.status = 500;
+        return {
+          friends: [],
+          incoming: [],
+          outgoing: [],
+        };
+      }
+    },
+    {
+      params: t.Object({
+        bapId: t.String(),
+      }),
+      response: FriendResponseSchema,
+      detail: friendEndpointDetail,
+    }
+  )
+  .get(
+    '/@/:bapId/messages',
+    async ({ params, query, set }) => {
+      try {
+        const response = await getDirectMessages({
+          bapId: params.bapId,
+          page: query.page ? Number.parseInt(query.page, 10) : 1,
+          limit: query.limit ? Number.parseInt(query.limit, 10) : 100,
+        });
+
+        Object.assign(set.headers, { 'Cache-Control': 'public, max-age=60' });
+        return response;
+      } catch (error) {
+        console.error('DM messages error:', error);
+        set.status = 500;
+        return {
+          bapID: params.bapId,
+          page: 1,
+          limit: 100,
+          count: 0,
+          results: [],
+          signers: [],
+        };
+      }
+    },
+    {
+      params: t.Object({ bapId: t.String() }),
+      query: MessageQuery,
+      response: DMResponseSchema,
+      detail: {
+        tags: ['social'],
+        description: 'Get encrypted direct messages for a BAP ID',
+        parameters: [
+          {
+            name: 'bapId',
+            in: 'path',
+            required: true,
+            schema: { type: 'string' },
+            description: 'Recipient BAP Identity Key',
+          },
+        ],
+      },
+    }
+  )
+  .get(
+    '/@/:bapId/messages/:targetBapId',
+    async ({ params, query, set }) => {
+      try {
+        const response = await getDirectMessages({
+          bapId: params.bapId,
+          targetBapId: params.targetBapId,
+          page: query.page ? Number.parseInt(query.page, 10) : 1,
+          limit: query.limit ? Number.parseInt(query.limit, 10) : 100,
+        });
+
+        Object.assign(set.headers, { 'Cache-Control': 'public, max-age=60' });
+        return response;
+      } catch (error) {
+        console.error('DM messages error:', error);
+        set.status = 500;
+        return {
+          bapID: params.bapId,
+          page: 1,
+          limit: 100,
+          count: 0,
+          results: [],
+          signers: [],
+        };
+      }
+    },
+    {
+      params: t.Object({ bapId: t.String(), targetBapId: t.String() }),
+      query: MessageQuery,
+      response: DMResponseSchema,
+      detail: {
+        tags: ['social'],
+        description: 'Get encrypted direct messages between two BAP IDs',
+        parameters: [
+          {
+            name: 'bapId',
+            in: 'path',
+            required: true,
+            schema: { type: 'string' },
+            description: 'Recipient BAP Identity Key',
+          },
+          {
+            name: 'targetBapId',
+            in: 'path',
+            required: true,
+            schema: { type: 'string' },
+            description: 'Target BAP Identity Key',
+          },
+        ],
+      },
+    }
+  )
+  .ws('/@/:bapId/messages/:targetBapId/listen', {
+    body: MessageListenParams,
+    open: async (ws) => {
+      const { bapId, targetBapId } = ws.data.params;
+      const identity = await fetchBapIdentityData(bapId);
+      if (!identity?.currentAddress) {
+        throw new Error('Invalid BAP identity');
+      }
+      const bapAddress = identity.currentAddress;
+      const targetIdentity = await fetchBapIdentityData(targetBapId);
+      if (!targetIdentity?.currentAddress) {
+        throw new Error('Invalid target BAP identity');
+      }
+      const targetAddress = targetIdentity.currentAddress;
+
+      const cursor = await watchDirectMessages({
+        bapId,
+        bapAddress,
+        targetBapId,
+        targetAddress,
+      });
+
+      cursor.on('change', (change: ChangeStreamInsertDocument<BmapTx>) => {
+        ws.send(change.fullDocument?._id);
+      });
+    },
+    detail: messageListenEndpointDetail,
+  })
+  .ws('/@/:bapId/messages/listen', {
+    body: MessageListenParams,
+    open: async (ws) => {
+      const { bapId } = ws.data.params;
+      const identity = await fetchBapIdentityData(bapId);
+      if (!identity?.currentAddress) {
+        throw new Error('Invalid BAP identity');
+      }
+      const bapAddress = identity.currentAddress;
+
+      const cursor = await watchAllMessages({
+        bapId,
+        bapAddress,
+      });
+
+      cursor.on('change', (change: ChangeStreamInsertDocument<BmapTx>) => {
+        ws.send({
+          tx: change.fullDocument?._id,
+          bap_id: change.fullDocument?.MAP?.[0]?.bapID,
+          algorithm_signing_component: change.fullDocument?.AIP?.[0]?.algorithm_signing_component,
+          aip_address: change.fullDocument?.AIP?.[0]?.address,
+        });
+      });
+    },
+    detail: messageListenEndpointDetail,
+  })
+  .get(
+    '/identities',
+    async ({ set }) => {
+      try {
+        console.log('=== Starting /identities request ===');
+
+        // Check Redis connection
+        console.log('Checking Redis connection...');
+        if (!client.isReady) {
+          console.error('Redis client is not ready');
+          set.status = 503;
+          return [];
+        }
+
+        // First try to get the cached identities list
+        const identitiesCacheKey = 'identities';
+        const cachedIdentities = await readFromRedis<CacheValue | CacheError>(identitiesCacheKey);
+
+        if (cachedIdentities?.type === 'identities' && Array.isArray(cachedIdentities.value)) {
+          console.log('Using cached identities list');
+          Object.assign(set.headers, {
+            'Cache-Control': 'public, max-age=60',
+          });
+          return cachedIdentities.value;
+        }
+
+        // If no cached list, get all signer-* keys from Redis
+        console.log('No cached identities list, checking individual signer caches');
+        const signerKeys = await client.keys('signer-*');
+        console.log(`Found ${signerKeys.length} cached signers`);
+
+        if (!signerKeys.length) {
+          console.log('No cached signers found');
+          return [];
+        }
+
+        // Get all cached signers
+        const identities = await Promise.all(
+          signerKeys.map(async (key) => {
+            try {
+              const cachedValue = await readFromRedis<CacheValue | CacheError>(key);
+              if (cachedValue?.type === 'signer' && 'value' in cachedValue) {
+                const identity = cachedValue.value;
+                if (identity && validateSignerData(identity).isValid) {
+                  return {
+                    idKey: identity.idKey || '',
+                    rootAddress: identity.rootAddress || '',
+                    currentAddress: identity.currentAddress || '',
+                    addresses: Array.isArray(identity.addresses)
+                      ? identity.addresses.map((addr) => ({
+                          address: addr.address || '',
+                          txId: addr.txId || '',
+                          block: typeof addr.block === 'number' ? addr.block : undefined,
+                        }))
+                      : [],
+                    identity:
+                      typeof identity.identity === 'string'
+                        ? identity.identity
+                        : typeof identity.identity === 'object'
+                          ? identity.identity
+                          : JSON.stringify(identity.identity || {}),
+                    identityTxId: identity.identityTxId || '',
+                    block: typeof identity.block === 'number' ? identity.block : 0,
+                    timestamp: typeof identity.timestamp === 'number' ? identity.timestamp : 0,
+                    valid: typeof identity.valid === 'boolean' ? identity.valid : true,
+                  };
+                }
+              }
+              return null;
+            } catch (error) {
+              console.error(`Error processing cached signer ${key}:`, error);
+              return null;
+            }
+          })
+        );
+
+        const filteredIdentities = identities.filter((id): id is BapIdentity => {
+          if (!id) return false;
+          return (
+            typeof id.idKey === 'string' &&
+            typeof id.rootAddress === 'string' &&
+            typeof id.currentAddress === 'string' &&
+            Array.isArray(id.addresses) &&
+            id.addresses.every(
+              (addr) =>
+                typeof addr.address === 'string' &&
+                typeof addr.txId === 'string' &&
+                (typeof addr.block === 'number' || addr.block === undefined)
+            ) &&
+            (typeof id.identity === 'string' || typeof id.identity === 'object') &&
+            typeof id.identityTxId === 'string' &&
+            typeof id.block === 'number' &&
+            typeof id.timestamp === 'number' &&
+            typeof id.valid === 'boolean'
+          );
+        });
+
+        console.log('\n=== Identity Processing Summary ===');
+        console.log('Total cached signers:', signerKeys.length);
+        console.log('Successfully processed:', filteredIdentities.length);
+        console.log('Failed/invalid:', signerKeys.length - filteredIdentities.length);
+
+        // Cache the filtered list
+        await saveToRedis<CacheValue>(identitiesCacheKey, {
+          type: 'identities',
+          value: filteredIdentities,
+        });
+
+        Object.assign(set.headers, {
+          'Cache-Control': 'public, max-age=60',
+        });
+        return filteredIdentities;
+      } catch (error: unknown) {
+        console.error('=== Error in /identities endpoint ===');
+        console.error('Error details:', error);
+        console.error('Stack trace:', error instanceof Error ? error.stack : 'No stack trace');
+        set.status = 500;
+        return [];
+      }
+    },
+    {
+      response: IdentityResponseSchema,
+      detail: {
+        tags: ['identities'],
+        description: 'Get all known BAP identities',
+        summary: 'List identities',
+        responses: {
+          200: {
+            description: 'List of BAP identities',
+            content: {
+              'application/json': {
+                schema: {
+                  type: 'array',
+                  items: {
+                    $ref: '#/components/schemas/BapIdentity',
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    }
+  );
